completed hw2_v4 (fixed double casting problem)

+++ hw2/hw2/PercolationStats.java
@@ -30,7 +30,7 @@ public class PercolationStats {
                 int col = StdRandom.uniform(0, N);
                 pfExp.open(row, col);
             }
-            threshold[t] = pfExp.numberOfOpenSites() / (N * N);
+            threshold[t] = (double) pfExp.numberOfOpenSites() / (N * N);
         }
     }
